refactor: refactored endgame state
