+++ src/components/HomeView/index.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import React, { useState, useCallback, useMemo } from 'react';
-import { Box, Typography, Button } from '@mui/material';
+import { Box } from '@mui/material';
 import { ScreenOpts } from '@/types';
 import {
   GAME_DIFF_EASY_INTERVAL,
@@ -13,6 +13,7 @@ import {
 } from '@/cfg/game';
 import { useSettings } from '@/hooks/useSettings';
 import { HomeWelcome } from './welcome';
+import { Endgame } from './endgame';
 import { GameView } from '../GameView';
 
 export const HomeView = () => {
@@ -61,22 +62,7 @@ export const HomeView = () => {
           screenSet={screenSet}
         />
       )}
-      {screen === 'end-game' && (
-        <Box>
-          <Box textAlign={'center'}>
-            <Typography variant="h3">Your Score: {score}</Typography>
-          </Box>
-          <Box textAlign={'center'} mt={3}>
-            <Button
-              size="large"
-              variant="contained"
-              onClick={() => screenSet('no-game')}
-            >
-              EXIT
-            </Button>
-          </Box>
-        </Box>
-      )}
+      {screen === 'end-game' && <Endgame screenSet={screenSet} score={score} />}
     </Box>
   );
 };
