+++ src/components/Layout/index.tsx
@@ -14,7 +14,7 @@ interface LayoutProps {
   children: React.ReactNode;
 }
 
-const VERSION = '1.1.0';
+const VERSION = '1.2.0';
 
 export const Layout = ({ children }: LayoutProps) => {
   return (
